crm/backend/requirements.txt
asgiref==3.8.1
Django==5.2.3
django-allauth==65.9.0     # Handles email sending + token generation
django-cors-headers==4.7.0
django-magiclink==1.3.0
djangorestframework==3.16.0
packaging==21.3
pyparsing==3.2.3
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
